+++ plugins/nf-quilt/src/main/nextflow/quilt/QuiltProduct.groovy
@@ -69,11 +69,7 @@ class QuiltProduct {
         String dir = pkg.packageDest()
         Path path  = Paths.get(dir, filename)
         try {
-            if (Files.exists(path)) {
-                path << text.bytes
-            } else {
                 Files.write(path, text.bytes)
-            }
         }
         catch (Exception e) {
             log.error("writeString: cannot write `$text` to `$path` for `${pkg}`")